+++ bin/bwa_mem.pl
@@ -2,7 +2,7 @@
 
 ##########LICENCE##########
 # PCAP - NGS reference implementations and helper code for the ICGC/TCGA Pan-Cancer Analysis Project
-# Copyright (C) 2014-2017 ICGC PanCancer Project
+# Copyright (C) 2014-2018 ICGC PanCancer Project
 #
 # This program is free software; you can redistribute it and/or
 # modify it under the terms of the GNU General Public License
@@ -87,7 +87,12 @@ sub cleanup {
 }
 
 sub setup {
-  my %opts;
+  my %opts = ('map_threads' => &PCAP::Bwa::bwa_mem_max_cores,
+              'mmqcfrac' => 0.05,
+              'threads' => 1,
+              'csi' => undef,
+             );
+
   GetOptions( 'h|help' => \$opts{'h'},
               'm|man' => \$opts{'m'},
               'v|version' => \$opts{'v'},
@@ -102,10 +107,13 @@ sub setup {
               'p|process=s' => \$opts{'process'},
               'i|index=i' => \$opts{'index'},
               'b|bwa=s' => \$opts{'bwa'},
+              'csi' => \$opts{'csi'},
               'c|cram' => \$opts{'cram'},
               'sc|scramble=s' => \$opts{'scramble'},
               'l|bwa_pl=s' => \$opts{'bwa_pl'},
               'g|groupinfo=s' => \$opts{'groupinfo'},
+              'q|mmqc' => \$opts{'mmqc'},
+              'qf|mmqcfrac:f' => \$opts{'mmqcfrac'},
   ) or pod2usage(2);
 
   pod2usage(-verbose => 1, -exitval => 0) if(defined $opts{'h'});
@@ -138,13 +146,10 @@ sub setup {
   delete $opts{'bwa'} unless(defined $opts{'bwa'});
   delete $opts{'scramble'} unless(defined $opts{'scramble'});
   delete $opts{'bwa_pl'} unless(defined $opts{'bwa_pl'});
-
-  $opts{'map_threads'} = &PCAP::Bwa::bwa_mem_max_cores unless(defined $opts{'map_threads'});
+  delete $opts{'mmqc'} unless(defined $opts{'mmqc'});
+  delete $opts{'csi'} unless(defined $opts{'csi'});
 
   PCAP::Cli::opt_requires_opts('scramble', \%opts, ['cram']);
-
-  # now safe to apply defaults
-  $opts{'threads'} = 1 unless(defined $opts{'threads'});
 
   my $tmpdir = File::Spec->catdir($opts{'outdir'}, 'tmpMap_'.$opts{'sample'});
   make_path($tmpdir) unless(-d $tmpdir);
@@ -209,15 +214,20 @@ bwa_mem.pl [options] [file(s)...]
 
   Optional parameters:
     -fragment    -f   Split input into fragements of X million repairs [10]
-    -nomarkdup   -n   Don't mark duplicates
-    -cram        -c   Output cram, see '-sc'
+    -nomarkdup   -n   Don't mark duplicates [flag]
+    -csi              Use CSI index instead of BAI for BAM files [flag].
+    -cram        -c   Output cram, see '-sc' [flag]
     -scramble    -sc  Single quoted string of parameters to pass to Scramble when '-c' used
                       - '-I,-O' are used internally and should not be provided
     -bwa         -b     Single quoted string of additional parameters to pass to BWA
                          - '-t,-p,-R' are used internally and should not be provided.
                          - '-v' is set to 1 unless '-bwa' is set.
-    -map_threads -mt  Number of cores applied to each parallel BWA job when '-t' exceeds this value and '-i' is not in use[6]
-    -groupinfo   -g   Readgroup information metadata file, values are not validated (yaml).
+    -map_threads -mt  Number of cores applied to each parallel BWA job when '-t' exceeds this value
+                      and '-i' is not in use [6]
+    -groupinfo   -g   Readgroup information metadata file, values are not validated (yaml) [file]
+    -mmqc        -q   Mark reads as QCFAIL (0x200, 512) if mismatch rate exceeded [flag]
+                       - Please see 'bwa_mem.pl -m'
+    -mmqcfrac    -qf  Mismatch fraction for -mmqc [0.05]
 
   Targeted processing:
     -process     -p   Only process this step then exit, optionally set -index
@@ -308,6 +318,10 @@ set this to a very large number or ensure that it is not changed.
 
 Disables duplicate marking, switching bammarkduplicates2 for bammerge.
 
+=item B<-csi>
+
+User CSI style index for final BAM file instead of default BAI.
+
 =item B<-cram>
 
 Final output file will be a CRAM file instead of BAM.  To tune the the compression methods see then
@@ -342,6 +356,23 @@ Recommend leaving this as the default and using increments of 6 for '-threads'.
 Readgroup information metadata file, please see the PCAP wiki for format:
 
 https://github.com/cancerit/PCAP-core/wiki/File-Formats-groupinfo.yaml
+
+=item B<-mmqc>
+
+Mark reads as QCFAIL (0x200, 512) using the mismatchQc program, also adds aux tag 'mm:A:Y'.
+
+WARNING:
+bwa_mem.pl will exclude all QCFAIL reads from mapping. If a BAM/CRAM file has been created using
+this option please ensure that you pre-process the file to remove the flag 512 if you intend to
+reprocess based on that output.
+
+e.g.
+
+bammaskflags maskneg=512 auxexists=mm < mmqc.bam > cleaned.bam
+
+=item B<-mmqcfrac>
+
+Mismatch fraction to pass through to mismatchQc
 
 =back